requirements.txt
boto==2.49.0
boto3==1.9.253
botocore==1.12.253
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.6
django-cors-headers==3.1.1
django-rest-framework==0.1.0
django-storages==1.7.2
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
docutils==0.15.2
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==6.2.0
psycopg2==2.8.3
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
